nptl: sh: Fix Race conditions in pthread cancellation (BZ#12683)
This patch adds the sh modifications required for the BZ#12683.
It basically adds the required ucontext_get_pc function and a
cancellable syscall wrapper.

SH requires an arch-specific syscall_cancel because the
INTERNAL_SYSCALL_NCS adds the required or instruction to workaround
a hardware bug [1].  The implementation were based on on default C
version built with GCC 6.2.1.

Checked against a build and make check run-built-tests=no for
sh4-linux-gnu.

	* sysdeps/unix/sysv/linux/sh/syscall_cancel.S: New file.
	* sysdeps/unix/sysv/linux/sh/sigcontextinfo.h (ucontext_get_pc):
	New function.

[1] http://documentation.renesas.com/eng/products/mpumcu/tu/tnsh7456ae.pdf

Signed-off-by: Adhemerval Zanella <adhemerval.zanella@linaro.org>